requirements.txt
Django==2.1.7
djangorestframework==3.9.2
Markdown==3.0.1
mysqlclient==1.4.2.post1
pkg-resources==0.0.0
pytz==2018.9
